+++ packages/snap-controller/src/Finder/FinderController.ts
@@ -25,8 +25,6 @@ export class FinderController extends AbstractController {
 		this.config = deepmerge(defaultConfig, this.config);
 		this.store.setConfig(this.config);
 
-		this.urlManager = this.urlManager;
-
 		// set the root URL on urlManager
 		if (this.config.url) {
 			this.urlManager = this.urlManager.withConfig((translatorConfig) => {
@@ -73,9 +71,11 @@ export class FinderController extends AbstractController {
 	};
 
 	reset = (): void => {
-		this.urlManager = this.urlManager.remove('filter');
-		this.store.storage.clear();
-		this.search();
+		// only need to reset when selections have been made
+		if (this.urlManager.state.filter) {
+			this.store.storage.clear();
+			this.urlManager.remove('filter').go();
+		}
 	};
 
 	search = async (): Promise<void> => {