+++ packages/snap-preact-components/src/components/Atoms/Image/Image.tsx
@@ -16,9 +16,10 @@ const CSS = {
 			display: 'flex',
 			flexDirection: 'column',
 			justifyContent: 'center',
-			height: '100%',
+			height: 'auto',
 			'& img': {
 				visibility,
+				flexShrink: '0',
 				objectFit: 'contain',
 				maxWidth: '100%',
 				maxHeight: '100%',
@@ -32,6 +33,7 @@ export function Image(properties: ImageProps): JSX.Element {
 	const props: ImageProps = {
 		// default props
 		fallback: FALLBACK_IMAGE_URL,
+		lazy: true,
 		// global theme
 		...globalTheme?.components?.image,
 		// props
@@ -39,7 +41,7 @@ export function Image(properties: ImageProps): JSX.Element {
 		...properties.theme?.components?.image,
 	};
 
-	const { alt, src, fallback, hoverSrc, onMouseOver, onMouseOut, onLoad, onClick, disableStyles, className, style } = props;
+	const { alt, src, fallback, hoverSrc, lazy, onMouseOver, onMouseOut, onError, onLoad, onClick, disableStyles, className, style } = props;
 
 	const [visibility, setVisibility] = useState('hidden');
 	const [isHovering, setHover] = useState(false);
@@ -65,13 +67,16 @@ export function Image(properties: ImageProps): JSX.Element {
 					src={(isHovering ? hoverSrc : src) || fallback}
 					alt={alt}
 					title={alt}
-					loading="lazy"
-					onLoad={() => {
+					loading={lazy ? 'lazy' : undefined}
+					onLoad={(e) => {
 						setVisibility('visible');
-						onLoad && onLoad();
+						onLoad && onLoad(e as any);
 					}}
 					onClick={(e) => onClick && onClick(e as any)}
-					onError={(e) => ((e.target as HTMLImageElement).src = fallback)}
+					onError={(e) => {
+						(e.target as HTMLImageElement).src = fallback;
+						onError && onError(e as any);
+					}}
 					onMouseOver={(e) => {
 						hoverSrc && setHover(true);
 						onMouseOver && onMouseOver(e as any);
@@ -93,6 +98,8 @@ export interface ImageProps extends ComponentProps {
 	hoverSrc?: string;
 	onMouseOver?: (e: MouseEvent) => void;
 	onMouseOut?: (e: MouseEvent) => void;
-	onLoad?: () => void;
+	onError?: (e: Event) => void;
+	onLoad?: (e: Event) => void;
 	onClick?: (e: MouseEvent) => void;
+	lazy?: boolean;
 }