+++ packages/snap-preact-demo/webpack.universal.js
@@ -1,13 +1,16 @@
 const { merge } = require('webpack-merge');
 const common = require('./webpack.common.js');
 const path = require('path');
+const childProcess = require('child_process');
+const branchName = childProcess.execSync('git rev-parse --abbrev-ref HEAD').toString().trim();
 
 module.exports = merge(common, {
 	mode: 'production',
 	entry: './src/universal.js',
 	output: {
-		filename: 'bundle.js',
-		chunkFilename: 'bundle.chunk.[fullhash:8].[id].js',
+		filename: 'universal.bundle.js',
+		chunkFilename: 'universal.bundle.chunk.[fullhash:8].[id].js',
+		chunkLoadingGlobal: `${branchName}BundleChunks`,
 	},
 	target: 'browserslist:universal',
 	module: {